Rugby and new Shuttles

+++ src/main/java/com/hillel/gorich/lessons/home_work7/Rugby.java
@@ -8,8 +8,8 @@ public class Rugby {
         int max = 40;
         int[] team1 = new int[25];
         int[] team2 = new int[25];
-        double sumAgeTeam1 = 0;
-        double sumAgeTeam2 = 0;
+        int sumAgeTeam1 = 0;
+        int sumAgeTeam2 = 0;
         double averageAgeTeam1 = 0;
         double averageAgeTeam2 = 0;
 
@@ -48,14 +48,14 @@ public class Rugby {
         for (int i = 0; i < team1.length; i++) {
             sumAgeTeam1 += team1[i];
         }
-        averageAgeTeam1 = sumAgeTeam1 / team1.length;
+        averageAgeTeam1 = ((double) sumAgeTeam1) / team1.length;
         System.out.println("Average age of first team players: " + Math.round(averageAgeTeam1));
         System.out.println();
 
         for (int i = 0; i < team2.length; i++) {
             sumAgeTeam2 += team2[i];
         }
-        averageAgeTeam2 = sumAgeTeam2 / team2.length;
+        averageAgeTeam2 = ((double) sumAgeTeam2) / team2.length;
         System.out.println("Average age of second team players: " + Math.round(averageAgeTeam2));
         System.out.println();
